refactor(io): dedup FileUtils::copyPandaDeps into Serialization::copyPandaFile
The two functions independently walked a .panda file's fileBackedICs
metadata to recursively copy IC dependencies — copyPandaFile (used by
"Add IC from file") also copied the top-level file itself and used
readPreamble()'s try/catch alone, while copyPandaDeps (used by
WorkSpace::save()'s directory-change copy step) skipped the top-level
copy and additionally pre-checked the magic header/app-name before
parsing, avoiding readPreamble()'s exception path on the common case
of non-panda content encountered while walking dependency blobs.

Ported that pre-check into copyPandaFile so the merged implementation
keeps every defensive property either version had, then removed
copyPandaDeps and pointed its one caller (WorkSpace::save()) at
copyPandaFile directly. Migrated its dedicated regression coverage
(including the circular-fileBackedICs-metadata cycle-guard test) to
TestSerialization, which previously had no unit-level test of
copyPandaFile at all — only the libFuzzer harness exercised it. Also
dropped FuzzCopyPanda's now-redundant third phase that called the
removed function directly.
