+++ Tests/Unit/Serialization/TestFileUtils.cpp
@@ -3,13 +3,10 @@
 
 #include "Tests/Unit/Serialization/TestFileUtils.h"
 
-#include <QDataStream>
 #include <QFile>
-#include <QRectF>
 #include <QTemporaryDir>
 
 #include "App/IO/FileUtils.h"
-#include "App/IO/Serialization.h"
 #include "Tests/Common/TestUtils.h"
 
 void TestFileUtils::testCopyToDirEmptyPath()
@@ -66,53 +63,6 @@ void TestFileUtils::testCopyToDirResourcePath()
     QVERIFY(true); // no crash
 }
 
-void TestFileUtils::testCopyPandaDepsBasic()
-{
-    // copyPandaDeps takes 3 args: pandaPath, srcDir, destDir
-    QTemporaryDir sourceDir;
-    QTemporaryDir destDir;
-
-    // Write a minimal file that won't parse as a real .panda — just verifies no crash
-    QString sourcePanda = sourceDir.path() + "/circuit.panda";
-    QFile pandaFile(sourcePanda);
-    QVERIFY(pandaFile.open(QIODevice::WriteOnly));
-    pandaFile.write("not a real panda file");
-    pandaFile.close();
-
-    FileUtils::copyPandaDeps(sourcePanda, sourceDir.path(), destDir.path());
-    QVERIFY(true); // no crash
-}
-
-void TestFileUtils::testCopyPandaDepsRecursive()
-{
-    QTemporaryDir sourceDir;
-    QTemporaryDir destDir;
-
-    QString sourcePanda = sourceDir.path() + "/parent.panda";
-    QFile pandaFile(sourcePanda);
-    QVERIFY(pandaFile.open(QIODevice::WriteOnly));
-    pandaFile.write("not a real panda file");
-    pandaFile.close();
-
-    FileUtils::copyPandaDeps(sourcePanda, sourceDir.path(), destDir.path());
-    QVERIFY(true); // no crash
-}
-
-void TestFileUtils::testCopyPandaDepsNoDependencies()
-{
-    QTemporaryDir sourceDir;
-    QTemporaryDir destDir;
-
-    QString sourcePanda = sourceDir.path() + "/simple.panda";
-    QFile pandaFile(sourcePanda);
-    QVERIFY(pandaFile.open(QIODevice::WriteOnly));
-    pandaFile.write("not a real panda file");
-    pandaFile.close();
-
-    FileUtils::copyPandaDeps(sourcePanda, sourceDir.path(), destDir.path());
-    QVERIFY(true); // no crash
-}
-
 void TestFileUtils::testCopyToDirThrowsOnFailure()
 {
     // Pre-fix QFile::copy's bool return was discarded; a failed copy left the
@@ -129,41 +79,4 @@ void TestFileUtils::testCopyToDirThrowsOnFailure()
     // Destination directory does not exist — QFile::copy returns false.
     const QString badDest = sourceDir.path() + "/no/such/subdirectory";
     QVERIFY_THROWS(std::exception, FileUtils::copyToDir(sourceFile, badDest));
-}
-
-namespace {
-// Writes a minimal .panda file with only the header + a metadata map containing
-// a single fileBackedICs entry. Just enough for copyPandaDeps to traverse.
-void writePandaWithFileBackedIC(const QString &path, const QString &referencedIC)
-{
-    QFile out(path);
-    QVERIFY(out.open(QIODevice::WriteOnly));
-    QDataStream stream(&out);
-    Serialization::writePandaHeader(stream);
-
-    QMap<QString, QVariant> metadata;
-    metadata["dolphinFileName"] = QString();
-    metadata["sceneRect"] = QRectF();
-    metadata["fileBackedICs"] = QStringList{referencedIC};
-    stream << metadata;
-}
-} // namespace
-
-void TestFileUtils::testCopyPandaDepsTerminatesOnCircularMetadata()
-{
-    // Hand-craft two .panda files that reference each other in their
-    // fileBackedICs metadata. Pre-fix copyPandaDeps would infinite-recurse
-    // until stack overflow; the visited-set guard short-circuits the cycle.
-    QTemporaryDir sourceDir;
-    QTemporaryDir destDir;
-    QVERIFY(sourceDir.isValid() && destDir.isValid());
-
-    const QString aPath = sourceDir.path() + "/a.panda";
-    const QString bPath = sourceDir.path() + "/b.panda";
-    writePandaWithFileBackedIC(aPath, "b.panda");
-    writePandaWithFileBackedIC(bPath, "a.panda");
-
-    // Should return promptly — if it doesn't terminate the test runner kills us.
-    FileUtils::copyPandaDeps(aPath, sourceDir.path(), destDir.path());
-    QVERIFY(true);
 }
