+++ Tests/Unit/Serialization/TestFileUtils.h
@@ -18,14 +18,6 @@ private slots:
     void testCopyToDirFileAlreadyExists();
     void testCopyToDirResourcePath();
 
-    // copyPandaDeps tests (3 tests)
-    void testCopyPandaDepsBasic();
-    void testCopyPandaDepsRecursive();
-    void testCopyPandaDepsNoDependencies();
-
     // Regression: D17 — copy failures must throw
     void testCopyToDirThrowsOnFailure();
-
-    // Regression: D16 — circular fileBackedICs metadata must not infinite-recurse
-    void testCopyPandaDepsTerminatesOnCircularMetadata();
 };
